Fix bugs (replace TestContext when exist, init idLocationMap, separate Delete controller from Edit controller)
