+++ hit-core-api/src/main/java/gov/nist/hit/core/api/ResourceController.java
@@ -24,7 +24,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -277,40 +276,4 @@ public class ResourceController {
 		return res;
 	}
 	
-	@PreAuthorize("hasRole('deployer')")
-	@ApiOperation(value = "delete TestStep by id", nickname = "deleteTestStep")
-	@RequestMapping(value = "/cb/delete/testStep/{id}", method = RequestMethod.DELETE, produces = "application/json")
-	public ResourceUploadStatus deleteTestStep(@PathVariable("id") Long id) {
-		return resourceLoader.deleteTS(id);
-	}
-	
-	@PreAuthorize("hasRole('deployer')")
-	@ApiOperation(value = "delete TestCase by id", nickname = "deleteTestCase")
-	@RequestMapping(value = "/cb/delete/testCase/{id}", method = RequestMethod.DELETE, produces = "application/json")
-	public ResourceUploadStatus deleteTestCase(@PathVariable("id") Long id) {
-		return resourceLoader.deleteTC(id);
-	}
-	
-	@PreAuthorize("hasRole('deployer')")
-	@ApiOperation(value = "delete TestCaseGroup by id", nickname = "deleteTestCaseGroup")
-	@RequestMapping(value = "/cb/delete/testCaseGroup/{id}", method = RequestMethod.DELETE, produces = "application/json")
-	public ResourceUploadStatus deleteTestCaseGroup(@PathVariable("id") Long id) {
-		return resourceLoader.deleteTCG(id);
-	}
-	
-	@PreAuthorize("hasRole('deployer')")
-	@ApiOperation(value = "delete TestPlan by id", nickname = "deleteTestPlan")
-	@RequestMapping(value = "/cb/delete/testPlan/{id}", method = RequestMethod.DELETE, produces = "application/json")
-	public ResourceUploadStatus deleteTestPlan(@PathVariable("id") Long id) {
-		return resourceLoader.deleteTP(id);
-	}
-	
-	@PreAuthorize("hasRole('deployer')")
-	@ApiOperation(value = "delete CF TestCase by id", nickname = "deleteCFTestCase")
-	@RequestMapping(value = "/cf/delete/testCase/{id}", method = RequestMethod.DELETE, produces = "application/json")
-	public ResourceUploadStatus deleteCFTestCase(@PathVariable("id") Long id) {
-		return resourceLoader.deleteCFTC(id);
-	}
-
-
 }
