+++ hit-core-service/src/main/java/gov/nist/hit/core/service/ResourceLoader.java
@@ -387,6 +387,7 @@ public abstract class ResourceLoader extends ResourcebundleLoader {
 
 			if ((index = tmp.indexOf(tcs)) != -1) {
 				tcs.setId(tmp.get(index).getId());
+				tcs.getTestContext().setId(tmp.get(index).getTestContext().getId());
 				tmp.set(index, tcs);
 			} else {
 				tmp.add(tcs);
